requirements.txt
numpy
natsort
torch
torchvision
pillow